Actions (#136)
Actions can be now opened in a callout using `ctrl+space`. The actions are the same that
are also available via the bottons. This is a basic, working version, there's definitely stuff for
improvement, .e.g the positioning of the dropdown, sorting of the elements in the dropdown
and most probably also other actions could be added.

Other stuff:
- Introduce IAction (e.g. IButton is an IAction)
- Refactor/rename IActions
- Get rid of `void 0`
- Use first useReducer